# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ABCompareResult": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/ABGroup"
          },
          "b": {
            "$ref": "#/components/schemas/ABGroup"
          },
          "d": {
            "title": "D",
            "type": "number"
          },
          "d_thresh": {
            "title": "D Thresh",
            "type": "number"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "significant": {
            "title": "Significant",
            "type": "boolean"
          },
          "t": {
            "title": "T",
            "type": "number"
          },
          "t_thresh": {
            "title": "T Thresh",
            "type": "number"
          },
          "underpowered": {
            "title": "Underpowered",
            "type": "boolean"
          },
          "verdict_detail": {
            "title": "Verdict Detail",
            "type": "string"
          },
          "verdict_label": {
            "title": "Verdict Label",
            "type": "string"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          }
        },
        "required": [
          "metric",
          "a",
          "b",
          "t",
          "d",
          "significant",
          "underpowered",
          "verdict_label",
          "verdict_detail",
          "t_thresh",
          "d_thresh"
        ],
        "title": "ABCompareResult",
        "type": "object"
      },
      "ABGroup": {
        "properties": {
          "mean": {
            "title": "Mean",
            "type": "number"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "std": {
            "title": "Std",
            "type": "number"
          },
          "vals": {
            "items": {
              "type": "number"
            },
            "title": "Vals",
            "type": "array"
          }
        },
        "required": [
          "name",
          "mean",
          "std",
          "vals",
          "n"
        ],
        "title": "ABGroup",
        "type": "object"
      },
...
      "EdrLinks": {
        "properties": {
          "edr": {
            "items": {
              "type": "integer"
            },
            "title": "Edr",
            "type": "array"
          }
        },
        "required": [
          "edr"
        ],
        "title": "EdrLinks",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "links": {
            "$ref": "#/components/schemas/RunLinks"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "name",
          "seed",
          "data",
          "links"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunLinks": {
        "properties": {
          "articles": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Articles",
            "type": "array"
          },
          "edr": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Edr",
            "type": "array"
          }
        },
        "title": "RunLinks",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AGIseed Dashboard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/compare": {
      "get": {
        "operationId": "compare_api_runs_compare_get",
        "parameters": [
          {
            "description": "condition A (name)",
            "in": "query",
            "name": "a",
            "required": true,
            "schema": {
              "description": "condition A (name)",
              "title": "A",
              "type": "string"
            }
          },
          {
            "description": "condition B (name)",
            "in": "query",
            "name": "b",
            "required": true,
            "schema": {
              "description": "condition B (name)",
              "title": "B",
              "type": "string"
            }
          },
          {
            "description": "métrique numérique à comparer",
            "in": "query",
            "name": "metric",
            "required": true,
            "schema": {
              "description": "métrique numérique à comparer",
              "title": "Metric",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ABCompareResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/links": {
      "patch": {
        "description": "Associe une liste d'EDR à un run (store results/run_links.json, n'altère pas le run).",
        "operationId": "set_run_links_api_runs__run_id__links_patch",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EdrLinks"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Run Links Api Runs  Run Id  Links Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Run Links",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}